+++ src/uipath_langchain/agent/wrappers/job_attachment_wrapper.py
@@ -1,19 +1,23 @@
+from __future__ import annotations
+
 import json
-from typing import Any
+from typing import TYPE_CHECKING, Any
 
 from langchain_core.messages.tool import ToolCall
 from langchain_core.tools import BaseTool
 from langgraph.types import Command
 from pydantic import BaseModel
 
-from uipath_langchain.agent.react.job_attachments import (
+from uipath_langchain.agent.attachments.job_attachments import (
     get_job_attachment_paths,
     get_job_attachments,
     replace_job_attachment_ids,
 )
-from uipath_langchain.agent.react.json_utils import coerce_json_strings
-from uipath_langchain.agent.react.types import AgentGraphState
-from uipath_langchain.agent.tools.tool_node import AsyncToolWrapperWithState
+from uipath_langchain.agent.attachments.pydantic_json import coerce_json_strings
+
+if TYPE_CHECKING:
+    from uipath_langchain.agent.react.types import AgentGraphState
+    from uipath_langchain.agent.tools.tool_node import AsyncToolWrapperWithState
 
 
 def _parse(content: str) -> Any:
